Update plugin with minimal changes to support Zeek 3.1/3.2
Disable selectable file descriptor workaround on Zeek 3.1+ as it is not needed and does not work with the new event loop.
Import some more specific headers.
Unfortunately still using deprecated functions as it is difficult to conditionally use new namespaces due to them already being defined in the global namespace by the same header. We will need to change this for Zeek 4.1 either by adding conditional using statements based on the exact version deprecated namespaces are actually removed, or ask upstream to add another mechanism such as a macro definition.